feat(grasshopper): adds dataobject icons (#977)
* adds icon pngs

* changes icons and icon order
