+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Components/Objects/PropertyGroupPathsSelector.cs
@@ -23,7 +23,7 @@ public class PropertyGroupPathsSelector : ValueSet<IGH_Goo>
 
   public override Guid ComponentGuid => new Guid("8882BE3A-81F1-4416-B420-58D69E4CC8F1");
   protected override Bitmap Icon => Resources.speckle_inputs_property;
-  public override GH_Exposure Exposure => GH_Exposure.tertiary;
+  public override GH_Exposure Exposure => GH_Exposure.quarternary;
 
   protected override void LoadVolatileData()
   {
