+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Components/Objects/SpeckleGeometryPassthrough.cs
@@ -20,8 +20,8 @@ public class SpeckleGeometryPassthrough : GH_Component
     ) { }
 
   public override Guid ComponentGuid => GetType().GUID;
-  protected override Bitmap Icon => Resources.speckle_objects_object;
-  public override GH_Exposure Exposure => GH_Exposure.primary;
+  protected override Bitmap Icon => Resources.speckle_objects_geometry;
+  public override GH_Exposure Exposure => GH_Exposure.secondary;
 
   protected override void RegisterInputParams(GH_InputParamManager pManager)
   {
