+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Parameters/Wrappers/SpeckleBlockDefinitionWrapperParam.cs
@@ -35,7 +35,7 @@ public class SpeckleBlockDefinitionWrapperParam
 
   protected override Bitmap Icon => Resources.speckle_param_block_def;
 
-  public override GH_Exposure Exposure => GH_Exposure.secondary;
+  public override GH_Exposure Exposure => GH_Exposure.tertiary;
 
   public override void RegisterRemoteIDs(GH_GuidTable idList)
   {
