+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Properties/Resources.resx
@@ -151,9 +151,6 @@
   <data name="speckle_objects_filter" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\speckle_objects_filter.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
-  <data name="speckle_objects_object" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\speckle_objects_object.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
   <data name="speckle_objects_query" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\speckle_objects_query.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
@@ -178,9 +175,6 @@
   <data name="speckle_param_material" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\speckle_param_material.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
-  <data name="speckle_param_object" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\speckle_param_object.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
   <data name="speckle_param_properties" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\speckle_param_properties.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
@@ -198,5 +192,17 @@
   </data>
   <data name="speckle_param_block_instance" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\speckle_param_block_instance.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="speckle_objects_dataobject" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\speckle_objects_dataobject.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="speckle_objects_geometry" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\speckle_objects_geometry.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="speckle_param_dataobject" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\speckle_param_dataobject.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="speckle_param_geometry" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\speckle_param_geometry.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
 </root>
