+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Resources/speckle_objects_geometry.png
