+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Resources/speckle_param_geometry.png
